Move workflow CLI under workbench
