+++ npa/src/npa/cli/workbench/__init__.py
@@ -14,6 +14,7 @@ from npa.cli.isaac_lab import app as isaac_lab_app
 from npa.cli.workbench.sonic import app as sonic_app
 from npa.cli.workbench.lancedb import app as lancedb_app
 from npa.cli.workbench.detection_training import app as detection_training_app
+from npa.cli.workbench.workflow import app as workflow_app
 
 app = typer.Typer(
     name="workbench",
@@ -40,3 +41,4 @@ app.add_typer(isaac_lab_app, name="isaac-lab")
 app.add_typer(sonic_app, name="sonic")
 app.add_typer(lancedb_app, name="lancedb")
 app.add_typer(detection_training_app, name="detection-training")
+app.add_typer(workflow_app, name="workflow")
